+++ src/lib.rs
@@ -26,10 +26,10 @@ pub use shape::*;
 pub use vector::{Vector, vector};
 pub use world::{World, default_world};
 
-pub const EPSILON: f64 = 0.0001;
+pub const EPSILON: f32 = 0.0001;
 pub const REFLECTION_DEPTH: usize = 5;
 
 #[must_use]
-pub fn clamp(value: f64, min: f64, max: f64) -> f64 {
+pub fn clamp(value: f32, min: f32, max: f32) -> f32 {
     value.max(min).min(max)
 }
